TES-268: Enable rate limiting on signup endpoint
• Enable rate limiting in signup API route (3 requests/minute)
• Remove duplicate rate limiting from business logic handler
• Standardize rate limiting pattern across all auth endpoints
• Update tests to reflect architecture changes
• Maintain consistent error messages and analytics tracking
