+++ __tests__/api.signup.test.ts
@@ -1,12 +1,11 @@
-import { signupBusinessLogic } from '../lib/auth/signup-handler';
-import { rateLimitMap, SignupResponse, SignupSuccessResponse, SignupErrorResponse } from '../lib/auth/signup-handler';
+import { signupBusinessLogic, SignupResponse, SignupSuccessResponse, SignupErrorResponse } from '../lib/auth/signup-handler';
 
 describe('signupBusinessLogic', () => {
   let supabaseClient: any;
   let analytics: any;
 
   beforeEach(() => {
-    rateLimitMap.clear();
+    // Rate limiting is now handled at API route level
     supabaseClient = {
       auth: {
         signUp: jest.fn().mockResolvedValue({ error: null }),
@@ -18,7 +17,7 @@ describe('signupBusinessLogic', () => {
   });
 
   it('returns 200 OK for valid email/password', async () => {
-    const res = await signupBusinessLogic({ email: 'test@example.com', password: 'password123', ip: '1.1.1.1', supabaseClient, analytics });
+    const res = await signupBusinessLogic({ email: 'test@example.com', password: 'password123', supabaseClient, analytics });
     expect(res.status).toBe(200);
     expect((res.body as SignupSuccessResponse).status).toBe('ok');
     expect(analytics.capture).toHaveBeenCalledWith({ 
@@ -32,32 +31,22 @@ describe('signupBusinessLogic', () => {
   });
 
   it('returns 400 for invalid email', async () => {
-    const res = await signupBusinessLogic({ email: 'bad', password: 'password123', ip: '1.1.1.1', supabaseClient, analytics });
+    const res = await signupBusinessLogic({ email: 'bad', password: 'password123', supabaseClient, analytics });
     expect(res.status).toBe(400);
     expect((res.body as SignupErrorResponse).error).toBe('Invalid input');
   });
 
   it('returns 400 for short password', async () => {
-    const res = await signupBusinessLogic({ email: 'test@example.com', password: 'short', ip: '1.1.1.1', supabaseClient, analytics });
+    const res = await signupBusinessLogic({ email: 'test@example.com', password: 'short', supabaseClient, analytics });
     expect(res.status).toBe(400);
     expect((res.body as SignupErrorResponse).error).toBe('Invalid input');
   });
 
-  it('returns 429 for repeated attempts (rate limit)', async () => {
-    // Hit the endpoint 5 times to fill the rate limit
-    for (let i = 0; i < 5; i++) {
-      await signupBusinessLogic({ email: 'test2@example.com', password: 'password123', ip: '2.2.2.2', supabaseClient, analytics });
-    }
-    // 6th attempt should be rate limited
-    const res = await signupBusinessLogic({ email: 'test2@example.com', password: 'password123', ip: '2.2.2.2', supabaseClient, analytics });
-    expect(res.status).toBe(429);
-    expect((res.body as SignupErrorResponse).error).toBe('Too many sign-up attempts');
-    expect(analytics.capture).toHaveBeenCalledWith({ event: 'signup_rate_limited', properties: { ip: '2.2.2.2' } });
-  });
+  // Rate limiting test removed - now handled at API route level
 
   it('returns 400 for duplicate email (Supabase error)', async () => {
     supabaseClient.auth.signUp.mockResolvedValueOnce({ error: { message: 'Email already in use' } });
-    const res = await signupBusinessLogic({ email: 'dupe@example.com', password: 'password123', ip: '3.3.3.3', supabaseClient, analytics });
+    const res = await signupBusinessLogic({ email: 'dupe@example.com', password: 'password123', supabaseClient, analytics });
     expect(res.status).toBe(400);
     expect((res.body as SignupErrorResponse).error).toBe('Email already in use');
     expect(analytics.capture).toHaveBeenCalledWith({ event: 'signup_error', properties: { email: 'dupe@example.com', error: 'Email already in use' } });
